+++ react/countries.ts
@@ -2,6 +2,7 @@
 import defaultRules from './country/default'
 import ARE from './country/ARE'
 import ARG from './country/ARG'
+import AUS from './country/AUS'
 import BOL from './country/BOL'
 import BRA from './country/BRA'
 import CAN from './country/CAN'
@@ -16,6 +17,7 @@ import IDN from './country/IDN'
 import IND from './country/IND'
 import KOR from './country/KOR'
 import MEX from './country/MEX'
+import NZL from './country/NZL'
 import PER from './country/PER'
 import PRT from './country/PRT'
 import PRY from './country/PRY'
@@ -31,6 +33,7 @@ export default {
   defaultRules,
   ARE,
   ARG,
+  AUS,
   BOL,
   BRA,
   CAN,
@@ -45,6 +48,7 @@ export default {
   IND,
   KOR,
   MEX,
+  NZL,
   PER,
   PRT,
   PRY,